Factor out package loding deduplication
